Fix warning C4244 in src/layer/elu.cpp
C4244: '=': conversion from 'double' to 'float', possible loss of data

+++ src/layer/elu.cpp
@@ -47,7 +47,7 @@ int ELU::forward_inplace(Mat& bottom_top_blob, const Option& opt) const
         for (int i=0; i<size; i++)
         {
             if (ptr[i] < 0.f)
-                ptr[i] = alpha * (exp(ptr[i]) - 1.f);
+                ptr[i] = static_cast<float>(alpha * (exp(ptr[i]) - 1.f));
         }
     }
